The flag to indicate use of AES was not set with `wc_AesSetKeyDirect`.

```
$ ./configure --enable-aesctr --enable-aesni && make

# Before Change
$ ./wolfcrypt/benchmark/benchmark -aes-ctr
------------------------------------------------------------------------------
 wolfSSL version 4.5.0
------------------------------------------------------------------------------
wolfCrypt Benchmark (block bytes 1048576, min 1.0 sec each)
AES-128-CTR        240 MB took 1.003 seconds,  239.273 MB/s Cycles per byte =  12.34
AES-192-CTR        215 MB took 1.004 seconds,  214.196 MB/s Cycles per byte =  13.78
AES-256-CTR        190 MB took 1.017 seconds,  186.759 MB/s Cycles per byte =  15.81
Benchmark complete

# After Change
$ ./wolfcrypt/benchmark/benchmark -aes-ctr
------------------------------------------------------------------------------
 wolfSSL version 4.5.0
------------------------------------------------------------------------------
wolfCrypt Benchmark (block bytes 1048576, min 1.0 sec each)
AES-128-CTR        605 MB took 1.003 seconds,  603.231 MB/s Cycles per byte =   4.89
AES-192-CTR        565 MB took 1.008 seconds,  560.358 MB/s Cycles per byte =   5.27
AES-256-CTR        505 MB took 1.004 seconds,  502.904 MB/s Cycles per byte =   5.87
```

I also considered moving this block of code into `wc_AesSetKeyLocal`, but would need to investigate other possible impacts. @SparkiDev , what are your thoughts?