@dgarske 2020-09-09
Fixes #3284