@SparkiDev 2020-09-09
wc_AesSetKeyLocal is the C software implementation right now.
We could have wc_SetAesKey call wc_SetAesKeyDirect instead and pull the common parts out of wc_SetAesKey.